move package `library.plantuml` -> `library.plantuml.rules`
... to free up the namespace for further possible enhancements around PlantUML. Our upcoming `1.0.0` release is the perfect point in time to do this breaking change. This will allow us to also add some infrastructure to create diagrams in the future within the `library.plantuml` package and make it clear that all existing code has the context of creating rules from PlantUML.
