+++ archunit/src/main/java/com/tngtech/archunit/library/plantuml/rules/PlantUmlDiagram.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.tngtech.archunit.library.plantuml;
+package com.tngtech.archunit.library.plantuml.rules;
 
 import java.util.List;
 import java.util.Set;
