+++ archunit/src/test/java/com/tngtech/archunit/library/plantuml/rules/TestDiagram.java
@@ -1,4 +1,4 @@
-package com.tngtech.archunit.library.plantuml;
+package com.tngtech.archunit.library.plantuml.rules;
 
 import java.io.File;
 import java.io.IOException;
